requirements.txt
streamlit
openpyxl
streamlit-authenticator
pandas
python-docx
Pillow
psutil
